consumer/requirements.txt
confluent-kafka==2.5.3
mysql-connector-python==9.0.0
python-dotenv==1.0.1
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5